load_test_with_python_locust/requirements.txt
grpcio==1.31.0
grpcio-tools==1.31.0
locust==1.1.1
setuptools~=49.6.0
gevent~=20.6.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.2.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
geventhttpclient==1.4.5
google==3.0.0
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.2
protobuf==3.14.0
psutil==5.8.0
pyzmq==22.0.2
requests==2.25.1
six==1.15.0
soupsieve==2.1
urllib3==1.26.3
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
